Update all old methods
