+++ bcos-framework/bcos-framework/storage2/Storage.h
@@ -1,201 +1,153 @@
 #pragma once
-
-#include "StorageDef.h"
-#include <bcos-concepts/Basic.h>
-#include <bcos-concepts/ByteBuffer.h>
-#include <bcos-task/Coroutine.h>
-#include <bcos-task/Task.h>
-#include <bcos-task/Trait.h>
-#include <bcos-utilities/Ranges.h>
-#include <boost/container/small_vector.hpp>
-#include <boost/throw_exception.hpp>
-#include <functional>
+#include "bcos-task/Task.h"
+#include "bcos-task/Trait.h"
+#include "bcos-utilities/Ranges.h"
 #include <optional>
-#include <type_traits>
 
+// tag_invoke storage interface
 namespace bcos::storage2
 {
-
-template <class IteratorType>
-concept ReadIterator =
-    requires(IteratorType&& iterator) {
-        requires std::convertible_to<task::AwaitableReturnType<decltype(iterator.next())>, bool>;
-        requires std::same_as<typename task::AwaitableReturnType<decltype(iterator.key())>,
-            typename IteratorType::Key>;
-        requires std::same_as<typename task::AwaitableReturnType<decltype(iterator.value())>,
-            typename IteratorType::Value>;
-        requires std::convertible_to<task::AwaitableReturnType<decltype(iterator.hasValue())>,
-            bool>;
-    };
-
-template <class StorageType>
-concept ReadableStorage = requires(StorageType&& impl) {
-                              typename StorageType::Key;
-                              requires ReadIterator<task::AwaitableReturnType<decltype(impl.read(
-                                  std::declval<std::vector<typename StorageType::Key>>()))>>;
-                          };
-
-template <class StorageType>
-concept WriteableStorage =
-    requires(StorageType&& impl) {
-        typename StorageType::Key;
-        typename StorageType::Value;
-        requires task::IsAwaitable<decltype(impl.write(
-            std::vector<typename StorageType::Key>(), std::vector<typename StorageType::Value>()))>;
-    };
 
 struct STORAGE_BEGIN_TYPE
 {
 };
 inline constexpr STORAGE_BEGIN_TYPE STORAGE_BEGIN{};
 
-template <class StorageType>
-concept SeekableStorage =
-    requires(StorageType&& impl) {
-        typename StorageType::Key;
-        requires ReadIterator<task::AwaitableReturnType<decltype(impl.seek(
-            std::declval<typename StorageType::Key>()))>>;
-        requires ReadIterator<
-            task::AwaitableReturnType<decltype(impl.seek(std::declval<STORAGE_BEGIN_TYPE>()))>>;
-    };
+template <class Invoke>
+using ReturnType = typename task::AwaitableReturnType<Invoke>;
 
-template <class StorageType>
-concept ErasableStorage = requires(StorageType&& impl) {
-                              typename StorageType::Key;
-                              requires task::IsAwaitable<decltype(impl.remove(
-                                  std::declval<std::vector<typename StorageType::Key>>()))>;
-                          };
-
-template <storage2::ReadableStorage Storage>
-struct ReadIteratorTrait
+struct ReadSome
 {
-    using type = typename task::AwaitableReturnType<decltype(std::declval<Storage>().read(
-        std::declval<std::vector<typename Storage::Key>>()))>;
+    auto operator()(auto& storage, RANGES::input_range auto const& keys) const
+        -> task::Task<ReturnType<decltype(tag_invoke(*this, storage, keys))>>
+        requires RANGES::range<ReturnType<decltype(tag_invoke(*this, storage, keys))>>
+    {
+        co_return co_await tag_invoke(*this, storage, keys);
+    }
 };
-template <storage2::ReadableStorage Storage>
-using ReadIteratorType = typename ReadIteratorTrait<Storage>::type;
 
-template <storage2::SeekableStorage Storage>
-struct SeekIteratorTrait
+struct WriteSome
 {
-    using type = typename task::AwaitableReturnType<decltype(std::declval<Storage>().seek(
-        std::declval<typename Storage::Key>()))>;
+    auto operator()(auto& storage, RANGES::input_range auto&& keys,
+        RANGES::input_range auto&& values) const -> task::Task<void>
+        requires std::same_as<void,
+            ReturnType<decltype(tag_invoke(*this, storage, std::forward<decltype(keys)>(keys),
+                std::forward<decltype(values)>(values)))>>
+    {
+        co_await tag_invoke(*this, storage, std::forward<decltype(keys)>(keys),
+            std::forward<decltype(values)>(values));
+    }
 };
-template <storage2::SeekableStorage Storage>
-using SeekIteratorType = typename SeekIteratorTrait<Storage>::type;
 
-inline auto singleView(auto&& value)
+struct RemoveSome
 {
-    using ValueType = decltype(value);
-
-    if constexpr (std::is_rvalue_reference_v<ValueType>)
+    auto operator()(auto& storage, RANGES::input_range auto const& keys) const
+        -> task::Task<ReturnType<decltype(tag_invoke(*this, storage, keys))>>
+        requires std::same_as<void, ReturnType<decltype(tag_invoke(*this, storage, keys))>>
     {
-        return RANGES::views::single(std::forward<decltype(value)>(value));
+        co_await tag_invoke(*this, storage, keys);
     }
-    else if constexpr (std::is_const_v<ValueType>)
+};
+struct ReadOne
+{
+    auto operator()(auto& storage, auto const& key) const
+        -> task::Task<ReturnType<decltype(tag_invoke(*this, storage, key))>>
     {
-        return RANGES::views::single(std::cref(value)) |
-               RANGES::views::transform([](auto&& input) -> auto const& { return input.get(); });
+        co_return co_await tag_invoke(*this, storage, key);
     }
-    else
+};
+
+struct WriteOne
+{
+    auto operator()(auto& storage, auto&& key, auto&& value) const -> task::Task<void>
+        requires std::same_as<void,
+            ReturnType<decltype(tag_invoke(*this, storage, std::forward<decltype(key)>(key),
+                std::forward<decltype(value)>(value)))>>
     {
-        return RANGES::views::single(std::ref(value)) |
-               RANGES::views::transform([](auto&& input) -> auto& { return input.get(); });
+        co_await tag_invoke(
+            *this, storage, std::forward<decltype(key)>(key), std::forward<decltype(value)>(value));
     }
-}
+};
 
-inline task::Task<bool> existsOne(ReadableStorage auto& storage, auto const& key)
+struct RemoveOne
 {
-    auto it = co_await storage.read(storage2::singleView(key));
-    co_await it.next();
-    co_return co_await it.hasValue();
-}
+    auto operator()(auto& storage, auto const& key) const
+        -> task::Task<ReturnType<decltype(tag_invoke(*this, storage, key))>>
+        requires std::same_as<void, ReturnType<decltype(tag_invoke(*this, storage, key))>>
+    {
+        co_await tag_invoke(*this, storage, key);
+    }
+};
 
-inline task::Task<void> removeOne(ErasableStorage auto& storage, auto const& key)
+struct ExistsOne
 {
-    co_await storage.remove(storage2::singleView(key));
-    co_return;
-}
+    auto operator()(auto& storage, auto const& key) const
+        -> task::Task<ReturnType<decltype(tag_invoke(*this, storage, key))>>
+        requires std::same_as<bool, ReturnType<decltype(tag_invoke(*this, storage, key))>>
+    {
+        co_return co_await tag_invoke(*this, storage, key);
+    }
+};
 
-namespace detail
-{
-template <class FromStorage, class ToStorage>
-concept HasMemberMergeMethod =
-    requires(FromStorage& fromStorage, ToStorage& toStorage) {
-        requires SeekableStorage<FromStorage>;
-        requires task::IsAwaitable<decltype(toStorage.merge(fromStorage))>;
-    };
 struct Merge
 {
-    template <class ToStorage>
-        requires WriteableStorage<ToStorage> && ErasableStorage<ToStorage>
-    task::Task<void> operator()(SeekableStorage auto& fromStorage, ToStorage& toStorage) const
+    auto operator()(auto& fromStorage, auto& toStorage) const
+        -> task::Task<ReturnType<decltype(tag_invoke(*this, fromStorage, toStorage))>>
+        requires std::same_as<void, ReturnType<decltype(tag_invoke(*this, fromStorage, toStorage))>>
     {
-        if constexpr (HasMemberMergeMethod<std::remove_cvref_t<decltype(fromStorage)>,
-                          std::remove_cvref_t<decltype(toStorage)>>)
-        {
-            co_await toStorage.merge(fromStorage);
-        }
-        else
-        {
-            auto it = co_await fromStorage.seek(storage2::STORAGE_BEGIN);
-            while (co_await it.next())
-            {
-                auto&& key = co_await it.key();
-                if (co_await it.hasValue())
-                {
-                    co_await storage2::writeOne(toStorage, key, co_await it.value());
-                }
-                else
-                {
-                    co_await storage2::removeOne(toStorage, key);
-                }
-            }
-        }
+        co_await tag_invoke(*this, fromStorage, toStorage);
     }
 };
-}  // namespace detail
-constexpr inline detail::Merge merge{};
 
-template <ReadableStorage Storage>
-auto tag_invoke(bcos::storage2::tag_t<readSome> /*unused*/, Storage& storage,
-    RANGES::input_range auto const& keys)
-    -> task::Task<boost::container::small_vector<
-        std::optional<std::remove_cvref_t<
-            typename task::AwaitableReturnType<decltype(storage.read(keys))>::Value>>,
-        1>>
+/******************************************************************/
+// Basic methods without default implementation
+/******************************************************************/
+inline constexpr ReadSome readSome{};
+inline constexpr WriteSome writeSome{};
+inline constexpr RemoveSome removeSome{};
+
+/******************************************************************/
+// Advance methods with default implementation, depends on basic methods
+/******************************************************************/
+inline constexpr ReadOne readOne{};
+inline constexpr WriteOne writeOne{};
+inline constexpr RemoveOne removeOne{};
+inline constexpr ExistsOne existsOne{};
+inline constexpr Merge merge{};
+
+template <auto& Tag>
+using tag_t = std::decay_t<decltype(Tag)>;
+
+// Default implementations
+auto tag_invoke(bcos::storage2::tag_t<readOne> /*unused*/, auto& storage, auto const& key)
+    -> task::Task<std::remove_cvref_t<decltype(std::declval<
+        task::AwaitableReturnType<decltype(readSome(storage, RANGES::views::single(key)))>>()[0])>>
 {
-    using ValueType = std::remove_cvref_t<
-        typename task::AwaitableReturnType<decltype(storage.read(keys))>::Value>;
-    static_assert(std::is_copy_assignable_v<ValueType>);
-
-    boost::container::small_vector<std::optional<ValueType>, 1> values;
-    if constexpr (RANGES::sized_range<decltype(keys)>)
-    {
-        values.reserve(RANGES::size(keys));
-    }
-    auto it = co_await storage.read(keys);
-    while (co_await it.next())
-    {
-        if (co_await it.hasValue())
-        {
-            values.emplace_back(co_await it.value());
-        }
-        else
-        {
-            values.emplace_back();
-        }
-    }
-
-    co_return values;
+    auto values = co_await readSome(storage,
+        RANGES::views::single(std::cref(key)) |
+            RANGES::views::transform([](auto&& input) -> auto const& { return input.get(); }));
+    co_return std::move(values[0]);
 }
 
-task::Task<void> tag_invoke(bcos::storage2::tag_t<writeSome> /*unused*/,
-    WriteableStorage auto& storage, RANGES::input_range auto const& keys,
-    RANGES::input_range auto&& values)
+auto tag_invoke(bcos::storage2::tag_t<writeOne> /*unused*/, auto& storage, auto&& key, auto&& value)
+    -> task::Task<void>
 {
-    co_await storage.write(keys, std::forward<decltype(values)>(values));
-    co_return;
+    co_await writeSome(storage, RANGES::views::single(std::forward<decltype(key)>(key)),
+        RANGES::views::single(std::forward<decltype(value)>(value)));
+}
+
+auto tag_invoke(bcos::storage2::tag_t<removeOne> /*unused*/, auto& storage, auto const& key)
+    -> task::Task<void>
+{
+    co_await removeSome(storage, RANGES::views::single(std::cref(key)) | RANGES::views::transform([
+    ](auto&& input) -> auto const& { return input.get(); }));
+}
+
+auto tag_invoke(bcos::storage2::tag_t<existsOne> /*unused*/, auto& storage, auto const& key)
+    -> task::Task<bool>
+{
+    auto result = co_await readOne(storage, key);
+    co_return result.has_value();
 }
 
 }  // namespace bcos::storage2
