+++ bcos-framework/benchmark/benchmarkMemoryStorage.cpp
@@ -95,7 +95,7 @@ static void read(benchmark::State& state)
                 int i = (state.range(0) / state.threads()) * state.thread_index();
                 for (auto const& it : state)
                 {
-                    auto itAwaitable = co_await storage.read(storage2::singleView(
+                    auto itAwaitable = co_await storage.read(RANGES::views::single(
                         fixture.allKeys[(i + fixture.allKeys.size()) % fixture.allKeys.size()]));
                     co_await itAwaitable.next();
                     [[maybe_unused]] auto& value = co_await itAwaitable.value();
@@ -135,7 +135,7 @@ static void readAny(benchmark::State& state)
         int i = (state.range(0) / state.threads()) * state.thread_index();
         for (auto const& it : state)
         {
-            auto itAwaitable = co_await anyStorageView->read(storage2::singleView(
+            auto itAwaitable = co_await anyStorageView->read(RANGES::views::single(
                 fixture.allKeys[(i + fixture.allKeys.size()) % fixture.allKeys.size()]));
             co_await itAwaitable.next();
             [[maybe_unused]] auto value = co_await itAwaitable.value();
@@ -165,8 +165,8 @@ static void write(benchmark::State& state)
                 for (auto const& it : state)
                 {
                     auto index = (i + fixture.allKeys.size()) % fixture.allKeys.size();
-                    co_await storage.write(storage2::singleView(fixture.allKeys[index]),
-                        storage2::singleView(fixture.allValues[index]));
+                    co_await storage.write(RANGES::views::single(fixture.allKeys[index]),
+                        RANGES::views::single(fixture.allValues[index]));
                     ++i;
                 }
                 co_return;
